+++ tests/test_cli.py
@@ -8,7 +8,7 @@ from tests.testcase import HookeeTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2020, Alex Laird"
-__version__ = "1.2.0"
+__version__ = "1.2.2"
 
 
 class TestCli(HookeeTestCase):